+++ backend/.gitignore
@@ -1,1 +1,2 @@
 .env
+realtime_data.json
